Refactored more archive code into the service and updated documentation now that methods are public.
