+++ API.Tests/Services/Test Data/CoverImageTests/thumbnail.expected.jpg
