+++ API.Tests/Services/Test Data/CoverImageTests/thumbnail.jpg
