+++ API.Tests/Services/Test Data/CoverImageTests/v10 - nested folder.cbz
