+++ API.Tests/Services/Test Data/CoverImageTests/v10 - with folder.expected.jpg
