+++ API.Tests/Services/Test Data/CoverImageTests/v10.cbz
