+++ API/Interfaces/IArchiveService.cs
@@ -5,6 +5,8 @@ namespace API.Interfaces
     public interface IArchiveService
     {
         bool ArchiveNeedsFlattening(ZipArchive archive);
-        public void ExtractArchive(string archivePath, string extractPath);
+        void ExtractArchive(string archivePath, string extractPath);
+        int GetNumberOfPagesFromArchive(string archivePath);
+        byte[] GetCoverImage(string filepath, bool createThumbnail = false);
     }
 }
